# Optimization of the CP solver