--- a/src/CP/core/fixPoint.jl
+++ b/src/CP/core/fixPoint.jl
@@ -9,7 +9,7 @@ Return a tuple with a boolean corresponding to the feasibility and a `CPModifica
 - `new_constraints::Union{Array{Constraint}, Nothing}`: if this is filled with a set of constraints, 
 only those will be propagated in the first place.
 """
-function fixPoint!(model::CPModel, new_constraints::Union{Array{Constraint}, Nothing}=nothing, prunedDomains::Union{CPModification,Nothing}=nothing)
+function fixPoint!(model::CPModel, new_constraints::Union{Array{Constraint}, Nothing}=nothing, prunedDomains::Union{CPModification,Nothing}=nothing; isRbs::Bool=false)
     toPropagate = Set{Constraint}()
 
     if isnothing(prunedDomains)
@@ -39,7 +39,7 @@ function fixPoint!(model::CPModel, new_constraints::Union{Array{Constraint}, Not
     while !isempty(toPropagate)
         constraint = pop!(toPropagate)
         if !propagate!(constraint, toPropagate, prunedDomains)
-            triggerInfeasible!(constraint, model)   
+            triggerInfeasible!(constraint, model; isRbs=isRbs)   
             return false, prunedDomains
         end
     end