--- a/src/CP/core/search/dfs.jl
+++ b/src/CP/core/search/dfs.jl
@@ -34,7 +34,7 @@ function expandDfs!(toCall::Stack{Function}, model::CPModel, variableHeuristic::
     end
 
     # Fix-point algorithm
-    feasible, pruned = fixPoint!(model, newConstraints, prunedDomains)
+    feasible, pruned = fixPoint!(model, newConstraints, prunedDomains; isFailureBased=isa(variableHeuristic, FailureBasedVariableSelection))
     model.statistics.lastPruning=sum(map(x-> length(x[2]),collect(pruned)))
 
     if !feasible