--- a/src/CP/core/search/rbs.jl
+++ b/src/CP/core/search/rbs.jl
@@ -81,7 +81,7 @@ function expandRbs!(toCall::Stack{Function}, model::CPModel, nodeLimit::Int64, c
     end
 
     # Fix-point algorithm
-    feasible, pruned = fixPoint!(model, newConstraints, prunedDomains; isRbs=true)
+    feasible, pruned = fixPoint!(model, newConstraints, prunedDomains; isFailureBased=isa(variableHeuristic, FailureBasedVariableSelection))
     model.statistics.lastPruning=sum(map(x-> length(x[2]),collect(pruned)))
     
     if !feasible